pes ediyorum: long long double -> int

+++ src/sinirlar.c
@@ -6,10 +6,10 @@
 void sinirlar(FILE *fp)
 {
 	fprintf(fp,
-	        "Sayi araligi: %lld - %lld\n"
+	        "Sayi araligi: %d - %d\n"
 	        "Maksimum sayi miktari: %d\n"
 	        "Maksimum dosya ismi uzunlugu: %d\n",
-	        LLONG_MIN, LLONG_MAX,
+	        INT_MIN, INT_MAX,
 	        MAX_NUM,
 	        MAX_DOSYA_ISIM_UZUNLUGU - 1);
 }
